refactor(settings): extract Formatter / Clipper / Sources panels (#1600, part 3)
Three more SettingsDialog panels lifted out:

- FormatterSettings.svelte — the Refactor ▸ Format rule list + per-rule toggles +
  Restore-house-style. Self-contained (each toggle persists immediately via
  setFormatSettings). Moved verbatim.
- ClipperSettings.svelte — browser-clipper enable + pairing code + status.
  Self-contained; owns its own onMount(getState) load and per-change toggle.
- SourcesSettings.svelte — ingest options + the privileged-sites list. Its one
  Done-batched setting (importUpstreamTags) is `bind:`-bound back to the host,
  which still loads + saves it in handleDone (the AiSettings pattern), so
  behaviour is identical; SitesSettings moves inside it.

SettingsDialog sheds their state/logic/imports (1297 → 1093 lines). Pure
extraction — behaviour identical. pnpm lint clean; full suite green (5424).
